# apps/bubu-log/openapi.yaml
openapi: 3.1.0
info:
  title: Bubu Log API
  description: Baby activity tracking API
  version: 1.0.0
servers:
  - url: /api
    description: API server
paths:
...
  /activities/{id}:
    get:
      operationId: getActivity
      summary: Get single activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Activity details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '404':
          description: Activity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

    patch:
      operationId: updateActivity
      summary: Update activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActivityInput'
      responses:
        '200':
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      operationId: deleteActivity
      summary: Delete activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletion success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    ActivityType:
      type: string
      enum:
        - SLEEP
        - DIAPER
        - BREASTFEED
        - BOTTLE
        - PUMP
        - HEAD_LIFT
        - PASSIVE_EXERCISE
        - ROLL_OVER
        - PULL_TO_SIT
        - GAS_EXERCISE
        - BATH
        - OUTDOOR
        - EARLY_EDUCATION
        - SUPPLEMENT
        - SPIT_UP
    MilkSource:
      type: string
      enum:
        - BREAST_MILK
        - FORMULA
    PoopColor:
      type: string
      enum:
        - YELLOW
        - GREEN
        - BROWN
        - BLACK
        - WHITE
        - RED
    PeeAmount:
      type: string
      enum:
        - SMALL
        - MEDIUM
        - LARGE
    BreastFirmness:
      type: string
      enum:
        - SOFT    # 软
        - ELASTIC # 弹
        - HARD    # 硬
    SupplementType:
      type: string
      enum:
        - AD    # 维生素AD
        - D3    # 维生素D3
    SpitUpType:
      type: string
      enum:
        - NORMAL     # 普通吐奶
        - PROJECTILE # 喷射性吐奶
...
    Activity:
      type: object
      required:
        - id
        - type
        - startTime
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/ActivityType'
        startTime:
          type: string
          format: date-time
          description: 活动开始时间
        endTime:
          type: string
          format: date-time
          nullable: true
          description: 活动结束时间（正在进行的活动没有结束时间）
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        hasPoop:
          type: boolean
          nullable: true
        hasPee:
          type: boolean
          nullable: true
        poopColor:
          $ref: '#/components/schemas/PoopColor'
        poopPhotoUrl:
          type: string
          nullable: true
        peeAmount:
          $ref: '#/components/schemas/PeeAmount'
        burpSuccess:
          type: boolean
          nullable: true
        milkAmount:
          type: integer
          nullable: true
        milkSource:
          $ref: '#/components/schemas/MilkSource'
          nullable: true
        breastFirmness:
          $ref: '#/components/schemas/BreastFirmness'
          nullable: true
        supplementType:
          $ref: '#/components/schemas/SupplementType'
          nullable: true
        spitUpType:
          $ref: '#/components/schemas/SpitUpType'
          nullable: true
        count:
          type: integer
          nullable: true
          description: 次数（用于翻身、拉坐等计数类活动）
        notes:
          type: string
          nullable: true
...
    UpdateActivityInput:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ActivityType'
        startTime:
          type: string
          format: date-time
          description: 活动开始时间
        endTime:
          type: string
          format: date-time
          description: 活动结束时间
        hasPoop:
          type: boolean
        hasPee:
          type: boolean
        poopColor:
          $ref: '#/components/schemas/PoopColor'
        poopPhotoUrl:
          type: string
        peeAmount:
          $ref: '#/components/schemas/PeeAmount'
        burpSuccess:
          type: boolean
        milkAmount:
          type: integer
        milkSource:
          $ref: '#/components/schemas/MilkSource'
        breastFirmness:
          $ref: '#/components/schemas/BreastFirmness'
        supplementType:
          $ref: '#/components/schemas/SupplementType'
        spitUpType:
          $ref: '#/components/schemas/SpitUpType'
        count:
          type: integer
          description: 次数（用于翻身、拉坐等计数类活动）
        notes:
          type: string
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
  responses:
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'